+++ mayan/apps/document_states/links.py
@@ -155,6 +155,17 @@ link_workflow_template_transition_list = Link(
 
 # Workflow transition fields
 
+link_document_multiple_workflow_templates_launch = Link(
+    icon_class_path='mayan.apps.document_states.icons.icon_document_workflow_templates_launch',
+    text=_('Launch workflows'),
+    view='document_states:document_multiple_workflow_templates_launch',
+)
+link_document_single_workflow_templates_launch = Link(
+    args='resolved_object.pk',
+    icon_class_path='mayan.apps.document_states.icons.icon_document_workflow_templates_launch',
+    permissions=(permission_workflow_tools,), text=_('Launch workflows'),
+    view='document_states:document_single_workflow_templates_launch',
+)
 link_workflow_template_transition_field_create = Link(
     args='resolved_object.pk',
     icon_class_path='mayan.apps.document_states.icons.icon_workflow_transition_field',