+++ mayan/apps/document_states/views/workflow_template_views.py
@@ -2,19 +2,24 @@ from django.contrib import messages
 from django.db import transaction
 from django.template import RequestContext
 from django.urls import reverse_lazy
-from django.utils.translation import ugettext_lazy as _
+from django.utils.translation import ugettext_lazy as _, ungettext
 
+from mayan.apps.acls.models import AccessControlList
 from mayan.apps.documents.events import event_document_type_edited
-from mayan.apps.documents.models import DocumentType
+from mayan.apps.documents.models.document_models import Document
+from mayan.apps.documents.models.document_type_models import DocumentType
 from mayan.apps.documents.permissions import permission_document_type_edit
 from mayan.apps.views.generics import (
-    AddRemoveView, ConfirmView, SingleObjectCreateView, SingleObjectDeleteView,
-    SingleObjectDetailView, SingleObjectEditView, SingleObjectListView
+    AddRemoveView, ConfirmView, MultipleObjectFormActionView,
+    SingleObjectCreateView, SingleObjectDeleteView, SingleObjectDetailView,
+    SingleObjectEditView, SingleObjectListView
 )
 from mayan.apps.views.mixins import ExternalObjectMixin
 
 from ..events import event_workflow_edited
-from ..forms import WorkflowForm, WorkflowPreviewForm
+from ..forms import (
+    WorkflowForm, WorkflowMultipleSelectionForm, WorkflowPreviewForm
+)
 from ..icons import icon_workflow_template_list
 from ..links import link_workflow_template_create
 from ..models import Workflow
@@ -78,6 +83,67 @@ class DocumentTypeWorkflowTemplatesView(AddRemoveView):
                 obj.instances.filter(
                     document__document_type=self.main_object
                 ).delete()
+
+
+class DocumentWorkflowTemplatesLaunchView(MultipleObjectFormActionView):
+    form_class = WorkflowMultipleSelectionForm
+    model = Document
+    object_permission = permission_workflow_tools
+    pk_url_kwarg = 'document_id'
+    success_message = _('Workflows launched for %(count)d document')
+    success_message_plural = _('Workflows launched for %(count)d documents')
+
+    def get_extra_context(self):
+        result = {
+            'submit_label': _('Launch'),
+            'subtitle': _(
+                'Workflows already launched or workflows not applicable to '
+                'some documents when multiple documents are selected, '
+                'will be silently ignored.'
+            ),
+            'title': ungettext(
+                singular='Launch selected workflows for %(count)d document',
+                plural='Launch selected workflows for %(count)d documents',
+                number=self.object_list.count()
+            ) % {
+                'count': self.object_list.count(),
+            }
+        }
+
+        if self.object_list.count() == 1:
+            result.update(
+                {
+                    'object': self.object_list.first(),
+                    'title': _(
+                        'Launch selected workflows for document: %s'
+                    ) % self.object_list.first()
+                }
+            )
+
+        return result
+
+    def get_form_extra_kwargs(self):
+        workflows_union = Workflow.objects.filter(
+            document_types__in=self.object_list.values('document_type')
+        ).distinct()
+
+        result = {
+            'help_text': _('Workflows to be launched.'),
+            'permission': permission_workflow_tools,
+            'queryset': workflows_union,
+            'user': self.request.user
+        }
+
+        return result
+
+    def object_action(self, form, instance):
+        workflow_queryset = AccessControlList.objects.restrict_queryset(
+            permission=permission_workflow_tools,
+            queryset=form.cleaned_data['workflows'], user=self.request.user
+        )
+
+        for workflow in workflow_queryset:
+            workflow.launch_for(document=instance)
 
 
 class WorkflowTemplateCreateView(SingleObjectCreateView):